wip: config panel refresh
