+++ core/src/main/kotlin/nl/nlportal/core/frontend/service/FrontendConfigurationService.kt
@@ -13,12 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.nlportal.core.autoconfiguration
 
-import org.springframework.boot.context.properties.ConfigurationProperties
+package nl.nlportal.core.frontend.service
 
-@ConfigurationProperties(prefix = "nl-portal.config.theme", ignoreUnknownFields = true)
-data class CoreThemeConfiguration(
-    val style: String? = null,
-    val logo: String? = null,
-)
+import nl.nlportal.core.autoconfiguration.CoreThemeConfigurationProperties
+import org.springframework.cloud.context.config.annotation.RefreshScope
+import org.springframework.stereotype.Component
+import org.springframework.stereotype.Service
+
+@RefreshScope
+@Component
+class FrontendConfigurationService(
+    private var coreThemeConfigurationProperties: CoreThemeConfigurationProperties
+) {
+    fun getLogo(): String? = coreThemeConfigurationProperties.logo
+    fun getStyle(): String? = coreThemeConfigurationProperties.style
+}
