+++ core/src/test/kotlin/nl/nlportal/core/frontend/FrontendConfigurationResourceTest.kt
@@ -15,7 +15,7 @@
  */
 package nl.nlportal.core.nl.nlportal.core.frontend
 
-import nl.nlportal.core.autoconfiguration.CoreThemeConfiguration
+import nl.nlportal.core.autoconfiguration.CoreThemeConfigurationProperties
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
@@ -30,7 +30,7 @@ import java.nio.charset.Charset
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class FrontendConfigurationResourceTest(
     @Autowired private val webTestClient: WebTestClient,
-    @Autowired private val coreThemeConfiguration: CoreThemeConfiguration,
+    @Autowired private val coreThemeConfiguration: CoreThemeConfigurationProperties,
 ) {
     @Test
     fun `get style`() {
